Move calculators to app/api/api_v1/endpoints
